chore: move common code to a separate file
