+++ controllers/memcached_controller.go
@@ -29,7 +29,6 @@ import (
 	"k8s.io/utils/pointer"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
 	cortexv1alpha1 "github.com/opstrace/cortex-operator/api/v1alpha1"
 )
@@ -159,68 +158,6 @@ func (r *MemcachedReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&cortexv1alpha1.Cortex{}).
 		Complete(r)
-}
-
-type KubernetesResource struct {
-	obj     client.Object
-	ref     *corev1.LocalObjectReference
-	mutator controllerutil.MutateFn
-}
-
-type KubernetesResourceReconciler struct {
-	log    logr.Logger
-	client client.Client
-	cortex *cortexv1alpha1.Cortex
-	scheme *runtime.Scheme
-}
-
-func (krr *KubernetesResourceReconciler) Reconcile(
-	ctx context.Context,
-	r *KubernetesResource,
-) error {
-	// Set up garbage collection. The object (resource.obj) will be
-	// automatically deleted whent he owner (cortex) is deleted.
-	err := controllerutil.SetOwnerReference(krr.cortex, r.obj, krr.scheme)
-	if err != nil {
-		krr.log.Error(
-			err,
-			"failed to set owner reference on resource",
-			"kind", r.obj.GetObjectKind().GroupVersionKind().Kind,
-			"name", r.obj.GetName(),
-		)
-		return err
-	}
-
-	op, err := controllerutil.CreateOrUpdate(ctx, krr.client, r.obj, r.mutator)
-	if err != nil {
-		krr.log.Error(
-			err,
-			"failed to reconcile resource",
-			"kind", r.obj.GetObjectKind().GroupVersionKind().Kind,
-			"name", r.obj.GetName(),
-		)
-		return err
-	}
-
-	err = krr.client.Status().Update(ctx, krr.cortex)
-	if err != nil {
-		krr.log.Error(
-			err,
-			"failed to reconcile resource status",
-			"kind", r.obj.GetObjectKind().GroupVersionKind().Kind,
-			"name", r.obj.GetName(),
-		)
-		return err
-	}
-
-	krr.log.Info(
-		"Reconcile successful",
-		"operation", op,
-		"kind", r.obj.GetObjectKind().GroupVersionKind().Kind,
-		"name", r.obj.GetName(),
-	)
-
-	return nil
 }
 
 func NewMemcachedService(req ctrl.Request, name string) *KubernetesResource {
